fix(promql): pin native-histogram result column types so rate() decodes
The nine Histogram*Column outputs took their ClickHouse type from
whichever lowering built the projection: a bare selector and sum()
forwarded the OTel-CH columns verbatim (UInt64 / Array(UInt64)), while
rate()/increase() divided by the window and landed on Float64 /
Array(Float64). chclient.HistogramValue binds one set of scan
destinations by position, so the two shapes cannot both decode --
clickhouse-go/v2's Float64.ScanRow rejects a *uint64 destination
outright, which meant rate() over an exponential histogram failed at the
first row in production.

Pin the nine output types in chsql's emitter instead: counts and sums to
Float64, bucket ladders to Array(Float64), Scale and the two bucket
offsets to Int32. Float64 is the correct pin rather than the convenient
one -- rate() over a native histogram yields fractional counts, exactly
as reference Prometheus returns a *histogram.FloatHistogram.

Refs #1967

+++ internal/api/prom/histogram.go
@@ -74,7 +74,7 @@ func histogramFromValue(hv *chclient.HistogramValue) *Histogram {
 			Boundaries: boundariesLowerInclusiveUpperExclusive,
 			Lower:      formatWireFloat(-math.Pow(base, idx+1)),
 			Upper:      formatWireFloat(-math.Pow(base, idx)),
-			Count:      formatWireFloat(float64(count)),
+			Count:      formatWireFloat(count),
 		})
 	}
 	if hv.ZeroCount > 0 {
@@ -89,7 +89,7 @@ func histogramFromValue(hv *chclient.HistogramValue) *Histogram {
 			Boundaries: boundariesBothInclusive,
 			Lower:      formatWireFloat(lower),
 			Upper:      formatWireFloat(upper),
-			Count:      formatWireFloat(float64(hv.ZeroCount)),
+			Count:      formatWireFloat(hv.ZeroCount),
 		})
 	}
 	for i, count := range hv.PositiveBucketCounts {
@@ -101,12 +101,12 @@ func histogramFromValue(hv *chclient.HistogramValue) *Histogram {
 			Boundaries: boundariesLowerExclusiveUpperInclusive,
 			Lower:      formatWireFloat(math.Pow(base, idx)),
 			Upper:      formatWireFloat(math.Pow(base, idx+1)),
-			Count:      formatWireFloat(float64(count)),
+			Count:      formatWireFloat(count),
 		})
 	}
 
 	return &Histogram{
-		Count:   formatWireFloat(float64(hv.Count)),
+		Count:   formatWireFloat(hv.Count),
 		Sum:     formatWireFloat(hv.Sum),
 		Buckets: buckets,
 	}
